requirements.txt
aiogram==3.17.0
aiohttp==3.11.12
fastapi==0.115.8
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
psycopg2==2.9.10
